moved plots.py to experimental
